Extract JSONAPIPatchSource from JSONAPISource.
The jsonpatch extension warrants its own source, since it will typically
be used independently from the base JSONAPISource. This extraction
leads to a cleaner implementation of both sources.
