+++ Brocfile.js
@@ -22,7 +22,8 @@ var packages = [
               /(orbit\-common\/.+.js)/],
     exclude: [/orbit-common\/local-storage-source.js/,
               /orbit-common\/jsonapi-serializer.js/,
-              /orbit-common\/jsonapi-source.js/]
+              /orbit-common\/jsonapi-source.js/,
+              /orbit-common\/jsonapi-patch-source.js/]
   },
   {
     name: 'orbit-common-local-storage',
@@ -31,7 +32,8 @@ var packages = [
   {
     name: 'orbit-common-jsonapi',
     include: [/orbit-common\/jsonapi-serializer.js/,
-              /orbit-common\/jsonapi-source.js/]
+              /orbit-common\/jsonapi-source.js/,
+              /orbit-common\/jsonapi-patch-source.js/]
   }
 ];
