+++ lib/orbit-common/jsonapi-source.js
@@ -29,7 +29,6 @@ var JSONAPISource = Source.extend({
     this.namespace        = options.namespace || this.namespace;
     this.host             = options.host || this.host;
     this.headers          = options.headers || this.headers;
-    this.usePatch         = options.usePatch !== undefined ? options.usePatch : this.usePatch;
     this.SerializerClass  = options.SerializerClass || this.SerializerClass;
 
     // If `SerializerClass` is obtained through the _super chain, dereference
@@ -50,7 +49,6 @@ var JSONAPISource = Source.extend({
   host: null,
   headers: null,
   SerializerClass: JSONAPISerializer,
-  usePatch: false,
 
   /////////////////////////////////////////////////////////////////////////////
   // Transformable interface implementation
@@ -155,14 +153,6 @@ var JSONAPISource = Source.extend({
   /////////////////////////////////////////////////////////////////////////////
 
   _transformAdd: function(operation) {
-    if (this.usePatch) {
-      return this._transformAddWithPatch(operation);
-    } else {
-      return this._transformAddStd(operation);
-    }
-  },
-
-  _transformAddStd: function(operation) {
     var _this = this;
     var type = operation.path[0];
     var id = operation.path[1];
@@ -175,37 +165,7 @@ var JSONAPISource = Source.extend({
     );
   },
 
-  _transformAddWithPatch: function(operation) {
-    var _this = this;
-    var type = operation.path[0];
-    var id = operation.path[1];
-
-    var remoteOp = {
-      op: 'add',
-      path: '/-',
-      value: this.serializer.serializeRecord(type, operation.value)
-    };
-
-    return this.ajax(this.resourceURL(type), 'PATCH', {data: [ remoteOp ]}).then(
-      function(raw) {
-        if (raw && isArray(raw)) {
-          _this.deserialize(type, id, raw[0], operation);
-        } else {
-          _this._transformCache(operation);
-        }
-      }
-    );
-  },
-
   _transformReplace: function(operation) {
-    if (this.usePatch) {
-      return this._transformReplaceWithPatch(operation);
-    } else {
-      return this._transformReplaceStd(operation);
-    }
-  },
-
-  _transformReplaceStd: function(operation) {
     var _this = this;
     var type = operation.path[0];
     var id = operation.path[1];
@@ -225,38 +185,7 @@ var JSONAPISource = Source.extend({
     );
   },
 
-  _transformReplaceWithPatch: function(operation) {
-    var _this = this;
-    var type = operation.path[0];
-    var id = operation.path[1];
-    var value = operation.value;
-
-    var remoteOp = {
-      op: 'replace',
-      path: '/',
-      value: this.serializer.serializeRecord(type, value)
-    };
-
-    return this.ajax(this.resourceURL(type, id), 'PATCH', {data: [ remoteOp ]}).then(
-      function(raw) {
-        if (raw && isArray(raw)) {
-          _this.deserialize(type, id, raw[0], operation);
-        } else {
-          _this._transformCache(operation);
-        }
-      }
-    );
-  },
-
   _transformRemove: function(operation) {
-    if (this.usePatch) {
-      return this._transformRemoveWithPatch(operation);
-    } else {
-      return this._transformRemoveStd(operation);
-    }
-  },
-
-  _transformRemoveStd: function(operation) {
     var _this = this;
     var type = operation.path[0];
     var id = operation.path[1];
@@ -266,32 +195,7 @@ var JSONAPISource = Source.extend({
     });
   },
 
-  _transformRemoveWithPatch: function(operation) {
-    var _this = this;
-    var type = operation.path[0];
-    var id = operation.path[1];
-
-    var remoteOp = {
-      op: 'remove',
-      path: '/'
-    };
-
-    return this.ajax(this.resourceURL(type, id), 'PATCH', {data: [ remoteOp ]}).then(
-      function() {
-        _this._transformCache(operation);
-      }
-    );
-  },
-
   _transformAddLink: function(operation) {
-    if (this.usePatch) {
-      return this._transformAddLinkWithPatch(operation);
-    } else {
-      return this._transformAddLinkStd(operation);
-    }
-  },
-
-  _transformAddLinkStd: function(operation) {
     var _this = this;
 
     var type = operation.path[0];
@@ -315,48 +219,7 @@ var JSONAPISource = Source.extend({
     );
   },
 
-  _transformAddLinkWithPatch: function(operation) {
-    var _this = this;
-
-    var type = operation.path[0];
-    var id = operation.path[1];
-    var link = operation.path[3];
-    var relId = operation.path[4] || operation.value;
-    var linkDef = this.schema.linkDefinition(type, link);
-    var relType = linkDef.model;
-    var relResourceId = this.serializer.resourceId(relType, relId);
-    var remoteOp;
-
-    if (linkDef.type === 'hasMany') {
-      remoteOp = {
-        op: 'add',
-        path: '/-',
-        value: relResourceId
-      };
-    } else {
-      remoteOp = {
-        op: 'replace',
-        path: '/',
-        value: relResourceId
-      };
-    }
-
-    return this.ajax(this.resourceLinkURL(type, id, link), 'PATCH', {data: [ remoteOp ]}).then(
-      function() {
-        _this._transformCache(operation);
-      }
-    );
-  },
-
   _transformRemoveLink: function(operation) {
-    if (this.usePatch) {
-      return this._transformRemoveLinkWithPatch(operation);
-    } else {
-      return this._transformRemoveLinkStd(operation);
-    }
-  },
-
-  _transformRemoveLinkStd: function(operation) {
     var _this = this;
 
     var type = operation.path[0];
@@ -371,47 +234,7 @@ var JSONAPISource = Source.extend({
     );
   },
 
-  _transformRemoveLinkWithPatch: function(operation) {
-    var _this = this;
-
-    var type = operation.path[0];
-    var id = operation.path[1];
-    var link = operation.path[3];
-    var linkDef = this.schema.linkDefinition(type, link);
-    var remoteOp;
-
-    if (linkDef.type === 'hasMany') {
-      var relId = operation.path[4];
-      var relType = linkDef.model;
-      var relResourceId = this.serializer.resourceId(relType, relId);
-
-      remoteOp = {
-        op: 'remove',
-        path: '/' + relResourceId
-      };
-    } else {
-      remoteOp = {
-        op: 'remove',
-        path: '/'
-      };
-    }
-
-    return this.ajax(this.resourceLinkURL(type, id, link), 'PATCH', {data: [ remoteOp ]}).then(
-      function() {
-        _this._transformCache(operation);
-      }
-    );
-  },
-
   _transformReplaceLink: function(operation) {
-    if (this.usePatch) {
-      return this._transformReplaceLinkWithPatch(operation);
-    } else {
-      return this._transformReplaceLinkStd(operation);
-    }
-  },
-
-  _transformReplaceLinkStd: function(operation) {
     var _this = this;
 
     var type = operation.path[0];
@@ -440,46 +263,7 @@ var JSONAPISource = Source.extend({
     );
   },
 
-  _transformReplaceLinkWithPatch: function(operation) {
-    var _this = this;
-
-    var type = operation.path[0];
-    var id = operation.path[1];
-    var link = operation.path[3];
-    var relId = operation.path[4] || operation.value;
-
-    // Convert a map of ids to an array
-    if (isObject(relId)) {
-      relId = Object.keys(relId);
-    }
-
-    var linkDef = this.schema.linkDefinition(type, link);
-    var relType = linkDef.model;
-    var relResourceId = this.serializer.resourceId(relType, relId);
-    var remoteOp;
-
-    remoteOp = {
-      op: 'replace',
-      path: '/',
-      value: relResourceId
-    };
-
-    return this.ajax(this.resourceLinkURL(type, id, link), 'PATCH', {data: [ remoteOp ]}).then(
-      function() {
-        _this._transformCache(operation);
-      }
-    );
-  },
-
   _transformUpdateAttribute: function(operation) {
-    if (this.usePatch) {
-      return this._transformUpdateAttributeWithPatch(operation);
-    } else {
-      return this._transformUpdateAttributeStd(operation);
-    }
-  },
-
-  _transformUpdateAttributeStd: function(operation) {
     var _this = this;
     var type = operation.path[0];
     var id = operation.path[1];
@@ -497,25 +281,6 @@ var JSONAPISource = Source.extend({
 
     return this.ajax(this.resourceURL(type, id), 'PUT', {data: json}).then(
       function(raw) {
-        _this._transformCache(operation);
-      }
-    );
-  },
-
-  _transformUpdateAttributeWithPatch: function(operation) {
-    var _this = this;
-    var type = operation.path[0];
-    var id = operation.path[1];
-    var attr = operation.path[2];
-
-    var remoteOp = {
-      op: 'replace',
-      path: '/' + attr,
-      value: operation.value
-    };
-
-    return this.ajax(this.resourceURL(type, id), 'PATCH', {data: [ remoteOp ]}).then(
-      function() {
         _this._transformCache(operation);
       }
     );
@@ -632,14 +397,8 @@ var JSONAPISource = Source.extend({
       // console.log('ajax start', method, url);
 
       if (hash.data && method !== 'GET') {
-        // If contentType has not been specified, use the appropriate type
-        // according to the JSON API spec
         if (!hash.contentType) {
-          if (method === 'PATCH') {
-            hash.contentType = 'application/json-patch+json; charset=utf-8';
-          } else {
-            hash.contentType = 'application/vnd.api+json; charset=utf-8';
-          }
+          hash.contentType = _this.ajaxContentType(hash);
         }
         hash.data = JSON.stringify(hash.data);
       }
@@ -671,6 +430,10 @@ var JSONAPISource = Source.extend({
 
       Orbit.ajax(hash);
     });
+  },
+
+  ajaxContentType: function(url, method) {
+    return 'application/vnd.api+json; charset=utf-8';
   },
 
   ajaxHeaders: function() {
